integer() should return an int, oops

+++ src/Random.h
@@ -14,7 +14,7 @@ class Random
 			std::uniform_real_distribution<double> dist(min, max);
 			return dist(mt);
 		}
-		double integer(int min, int max)
+		int integer(int min, int max)
 		{
 			std::uniform_int_distribution<int> dist(min, max);
 			return dist(mt);
